requirements.txt
flask
gunicorn
openai
PyPDF2
pandas
python-dotenv
